Refactor: Provider 분리
