+++ modules/shared/provider/type.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface IProvider {
+  children: ReactNode;
+}
